Manually edited  the indenting from the last pull request
I don't comprehend why that Anon made his pull request with a different indenting at the beggining of each line, nor why he put the initial (original) part of the code (the part which let's you change the Colab's name) at the end of the notebook instead of at the beggining. Please enlightenment me if there is a practical reason for this.