Commit nº152: Merged most of the Beta branch (visual fixes only), v1.22
Only changes to the Beta were on correcting the Disclaimer table. Only thing left is to merge the all-in-one cell that I didn't touch for this version.
When that cell is merged, we will advance to either v1.23 or v2.0 depending on how big the changes that cell brings.
Also, in that same release we will change the name of BOTH this repository and the Colab to "Anons' Open-CYOAI Project" permanently or until a better name arrives. The fork devs are completly free to decide if they want to change of name or not.


Beware (looking at you visualfag) that this would mean we would have to change some reference links in the Colab code for it to work correctly (I hope that I myself don't forget of this).